+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file, following t
 - Selections support `instance_id`
 - Complex superposition extension allows RNA superposition
 - Option `mapSettings.defaultView` to set Volume Streaming view type
+- Added `.visual.sequenceColor`, `.visual.clearSequenceColor`
 
 ## [v3.6.0] - 2025-08-12
 - Option `hideCanvasControls` accepts "screenshot" value